Issues with next line parsing of \n

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: hshi-yun <hshi-yun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/18 17:22:23 by hshi-yun          #+#    #+#             */
-/*   Updated: 2024/08/25 17:05:46 by hshi-yun         ###   ########.fr       */
+/*   Updated: 2024/08/26 16:22:33 by hshi-yun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@
 
 char    *get_next_line(int fd);
 void    *ft_calloc(size_t n, size_t size);
-size_t	ft_strlen(const char *s);
-char	*ft_strjoin(char const *s1, char const *s2, size_t size);
+int     ft_strlen(const char *s);
+char	*ft_strjoin(char const *s1, char const *s2, int size);
 int		ft_strchr_index(const char *s, int c);
 void	trim_newline(char *s1, int position);
